Support disabling a cell
It would be useful to have a mechanism by which we could totally stop
scheduling to a particular cell or a group of cells by supporting the
concept of disabling cells. Given that we do not have any existing
means to disable a cell, this spec proposes a simple solution to
support this new feature in nova.

Change-Id: I22bfe4dbe8325e57cef9959c38adda94fef28d73
Implements: blueprint cell-disable