Express forbidden traits in placement API
The spec proposes new syntax and handling for expressing traits that
required to _not_ be present in the results of both GET
/resource_providers and GET /allocation_candidates. The proposed
syntax is designed to minimize a profusion of query parameters. But
it may be too weird, so a syntax using a new parameter is also
proposed.

blueprint placement-forbidden-traits

Change-Id: I4e25c0ae28cf7d9aa75ed75e164ae38d9f4eaa54