NUMA-aware live migration
When an instance with NUMA characteristics is live-migrated, those
characteristics are not recalculated on the destination compute host.
In the CPU pinning case, using the source host's pin mappings on the
destination can lead to multiple instances being pinned to the same
pCPUs. In the case of hugepage-backed instances, which are
NUMA-localized, an instance needs to have its NUMA mapping
recalculated on the destination compute host during a live migration.

Change-Id: I1c75c555821e9eb7c51deb7eee17e3977cf45f18