Spec for isolating configuration of placement database
This spec provides an optional configuration setting that allows the
placement database to be managed separately from the nova api database.
As described in the spec this will help with extraction of placement and
also provide some options for how people manage data when they want to
migrate.

This spec describes work that is already in progress, linked within.

blueprint optional-placement-database

Change-Id: I7c36e54828da5576cb4093deb49a5fdcd667b4e7