Add PENDING vm state
This feature adds support for the PENDING server state. Instead of
setting a server to the ERROR state when the Placement returns no
valid hosts for the requested server, the PENDING state is used if
the operator wishes so. This will allow the execution of subsequent
actions transparently to the end user.

Change-Id: Ie2989dee31a8d7242bf5b78019f016b3bc7e42a3
Implements: blueprint introduce-pending-vm-state