Enable rebuild for instances in cell0
Presently, it is allowed to rebuild servers in ERROR state, as long
as they have successfully started up before.

If a user tries to rebuild an instance in cell0, the request fails
with an exception of type InstanceInvalidState.

This spec summarizes the changes needed to enable the rebuilding of
instances that failed during creation.

Change-Id: I3cf92b987da014e8b01f8b1a3f774b03ce3bb0b7
Implements: blueprint enable-rebuild-for-instances-in-cell0