Add Generation to Consumers
Potential conflicts have been identified when more than one process
attempts to allocate resources for a given consumer, and there is
currently no way to detect these conflicts. We propose to add a
generation field to the consumers table in Placement, and implement the
same mechanism for tracking updates to consumers as is implemented for
ResourceProviders.

APIImpact

Blueprint: add-consumer-generation

Change-Id: I8128da7185e149ecde45c94bb00668ce3e949251