Allow having placement inventories with reserved value equal to total
Currently, we delete ironic node resource provider inventory from placement
to indicate that the node is not available for deployment during the cleaning
process or when in maintenance. This is not actually required as we can
reserve all the inventory, which describes more accurately what is going on,
and introduces issues like https://bugs.launchpad.net/nova/+bug/1771577.

APIImpact
blueprint: allow-reserved-equal-total-inventory
Change-Id: I0aec5e9f0a26430a9b104159c62ba1d07944ce9f