Add support for emulated virtual TPM
There are a class of applications which expect to use a TPM device
to store secrets.  In order to run these applications in a virtual
machine, it would be useful to expose a virtual TPM device within
the guest.  Accordingly, the suggestion is to create a placement
trait which could be requested in the flavor or image which would
cause such a device to be added to the VM by the relevant virt
driver.

Change-Id: I4b3a62b01edfbb0965f3366358735536718e6e6c
blueprint: add-emulated-virtual-tpm