Spec: Handling Reshaped Provider Trees
Virt drivers need to be able to change the structure of the provider
trees they expose. When moving existing resources, existing allocations
need to be moved along with the inventories. And this must be done in
such a way as to avoid races where a second entity can create or remove
allocations against the moving inventories.

Change-Id: I1508c8e12c75b24ec9da04468b700b60f055ec24
blueprint: reshape-provider-tree