Add a script for counting blueprints
This is useful for recording data during the cycle to use to create
a burndown chart for blueprints. The script counts: Targeted,
Approved, and Implemented blueprints. It also checks if there are any
approved specs whose corresponding blueprints need approval and emits
warnings for each one found.

Since this shares some code with the existing move_implemented_specs
script, the common code is factored out into an importable lib.py
module.

This also adds a tox target for running the script and makes it share
the same envdir as the move-implemented-specs target since both need
the same dependency of launchpadlib.

Change-Id: I9a9d3f7f2883a6eb151230da651d9a4c4fda77b4