Placement: any traits in allocation_candidate query
The GET /allocation_candidates request in Placement supports the
required query parameter. If the caller specifies a list of traits
in the required parameter then placement will limit the returned
allocation candidates to those RP trees that fulfill *every* traits in
that list. To support minimum bandwidth guarantees in Neutron + Nova we
need to be able to query allocation candidates that fulfill at least
one trait from a list of traits specified in the query. This is
required for the case when a Neutron network maps to more than one
physnet but the port's bandwidth request can be fulfilled from any
physnet the port's network maps to.

bp any-traits-in-allocation-candidates-query

Change-Id: I57c5611d7443070c176d213118857261b37eca0c