Placement: support mixing required traits with any traits
The previous spec I57c5611d7443070c176d213118857261b37eca0c proposed
to allow querying traits in the form of required=in:TRAIT1,TRAIT2.
This spec goes one step forward and propose to allow repeating the
required query parameter to support mixing both
required=TRAIT1,TRAIT2,!TRAIT3 and required=in:TRAIT1,TRAIT2 format
in a single query. This is needed for Neutron to be able to express
that a port needs a resource provider having a specific vnic_type
trait but also having one of the physnet traits the port's network maps
to.

bp mixing-required-traits-with-any-traits

Change-Id: I9695f0cb2c0263cd0e8ca552d3a9dd720690e466