Handling a down cell
This spec aims at addressing the behavioural changes that are required to
support some of the basic nova operations like listing of instances and
services when a cell goes down.

Note that this is just a re-proposal of the already approved version in
rocky, for stein. Initial half of the spec implementation was merged in
rocky and now the API changes required to complete the spec are under
review for merge in stein.

APIImpact

Previously-approved: Rocky

Change-Id: Ia7efedb2b2cd10555889d795f19847c1033b416d
Implements: blueprint handling-down-cell