VMware: add support for live migration
This is a proposal for adding support for live migration in the VMware
driver. When the VMware driver is used, each nova-compute is managing a
single vCenter cluster. For the purposes of this proposal we assume that
all nova-computes are managing clusters under the same vCenter server.
Migrating instances across vCenters is supported starting from vSphere
6.0. We can add support for this as a next-step after this blueprint is
implemented.

Previously-approved: Rocky

blueprint vmware-live-migration

Change-Id: I483a721f2a1101e4667c5baa20c71cfefa143a26