Per aggregate scheduling weight (spec)
Filter Scheduler uses the so-called weights during its work. A weigher
is a way to select the best suitable host from a group of valid hosts by
giving weights to all the hosts in the list.

In the current implementation, the weights are controlled by config
options like [filter_scheduler]/cpu_weight_multiplier.
The total weight of a compute node is calculated by combination of
several weighers:
weight = w1_multiplier * norm(w1) + w2_multiplier * norm(w2) + ...

As it is controlled by config options, the weights are global across the
whole deployment, and this is not efficient enough for operators.

This spec proposes to be able to override the weight configuration
options using aggregate metadata to provide a more fine-grained
control over resource weights.

Change-Id: I8270f3a51260c2e8fbbbb4f9c179729a28d138ef
Implements: blueprint per-aggregate-scheduling-weight