Detach and attach boot volumes - Stein
It is sometimes useful for a cloud user to be able to
detach and attach the boot volume of an instance that
is not running. Currently Nova does not allow this and
some operations assume it does not happen. This spec
proposes allowing the detach and attach of boot volumes
in the shelved_offloaded and powered_off states.

There is a previously approved spec in Mitaka, Newton and
Ocata, but it only includes the change for shelved instances,
this spec append that proposal by adding also the powered_off
state.

Co-Authored-By: Andrea Rosa <andrea.rosa@hpe.com>

blueprint detach-boot-volume

Change-Id: Ide993b5ee0dba08b0ed9d641f6b96b344bb14f27