Use conductor groups to partition nova-compute services for Ironic
This spec proposes using ironic's conductor group feature to limit the
subset of nodes which a nova-compute service will manage. This allows
for partitioning nova-compute services to a particular location
(building, aisle, rack, etc), and provides a way for operators to manage
the failure domain of a given nova-compute service.

Change-Id: I62b838ee5fc56d97fdd35d620dd76d50606d0f34
Implements: blueprint ironic-conductor-groups