Select cpu model from a list of cpu models
Currently we uses cpu_model in nova.conf (when cpu_mode is set to
‘custom’) to specify the CPU model the instance should use on this host.
This could have implications on availability of compute nodes for live
migration if you ended up with an "advanced" CPU model when all you
really cared about was an older feature flag.

To mitigate the problem , we use cpu_models to replace cpu_model, where
cpu_models is an ordered list of CPU models the host supports. The list
is ordered so that the more common and less advanced CPU models are
listed earlier. End users specify CPU features they required through
traits, libvirt driver then select the first one in the list that can
provide the required feature traits. This would make it more likely that
the instance could be live-migrated later on.

blueprint: cpu-model-selection
Change-Id: I546230ac115eb3de14eac25b9dd0c74aafdaccd5
Signed-off-by: Jack Ding <jack.ding@windriver.com>