Support server power state update through external event
This spec aims at providing more flexibility for operators regarding the
``_sync_power_states`` periodic task (which aligns the server states
between the database and the hypervisor) in nova with respect to use
cases for the baremetal instances (ironic). It proposes to make this
periodic power sync's "source of truth" configurable, depending on
situations, like to allow the physical instance to be the source of
truth and make nova update its database rather than enforcing the
database state onto the physical instance.

Story: 2004969
Task: 29422
Implements blueprint nova-support-instance-power-update
Change-Id: I91eaf14053ecac38dd116ec67feb1f5bafa64226