Boot a VM with an unaddressed port
Currently Nova only allows to boot a VM if all ports attached:
- Have an IP assigned.
- The IP allocation is deferred.

But Nova doesn't allow to spawn a VM with a port without an IP assigned.
However, some kind of applications (e.g.: routers) need a port which may
be receiving traffic for several addresses, already configured in other
devices.

At the moment, Neutron permits to create a port, assigned to a network
with the corresponding subnetworks and IP pools, without an IP.

bp boot-vm-with-unaddressed-port

Change-Id: I48c7ddab4dae201530231d9061f2ade0aa600c07