Re-propose cross-cell-resize spec for Train
There are a few clarifications added based on the stein
spec review feedback: https://review.openstack.org/#/c/616037/

A known issue is added for personality files not being retained
in the config drive during a cross-cell resize.

A link to the code series is added in the References section
and the History section is updated for Train.

Details about the need to hard-delete the instance and its
related records is also added since the Stein spec.

Previously-approved: Stein

Spec for blueprint cross-cell-resize

Change-Id: I68498afd481f7291a6102928d7999b4be49ded7a