Spec: Use in_tree getting allocation candidates
In Stein, we introduced in_tree=<rp_uuid> parameter in placement for
the ``GET /allocation_candidates`` endpoints, which limits the
response to resource providers within the same tree of the specified
resource provider.

This spec proposes to use this parameter for optimization when we move
instances and the target host is already picked before asking to the
scheduler.

Change-Id: I99064bf9497c0ede6c9bada2fbc783204364f6f2
Blueprint: use-placement-in-tree