Enable rebuild for instances in cell0
Presently, it is allowed to rebuild servers in ERROR state, as long
as they have successfully started up before.

If a user tries to rebuild an instance in cell0, the request fails
with an exception of type InstanceInvalidState.

This spec summarizes the changes needed to enable the rebuilding of
instances that failed during creation.

Previously-approved: Stein
Change-Id: I86ff95a3de63979c453af271c4dcd8bd0e80ca3b
Implements: blueprint enable-rebuild-for-instances-in-cell0