Add PENDING vm state
This feature adds support for the PENDING server state. Instead of
setting a server to the ERROR state when the Placement returns no
valid hosts for the requested server, the PENDING state is used if
the operator wishes so. This will allow the execution of subsequent
actions transparently to the end user.

Previously-approved: Stein
Change-Id: Ia605b2a64a468be264be86f27d46ce00e6a16b10
Implements: blueprint introduce-pending-vm-state