Update SEV work item to new approach based on MEM_ENCRYPTION_CONTEXT
Between the Stein and Train versions of the AMD SEV spec, the design
pivoted from using a trait to using a resource class.  However one
work item was missed from the switch, so fix it.

The significant point here is that the original RequestSpec doesn't
necessarily contain the SEV trait, e.g. if the user or operator has no
preference whether to boot the VM using SEV or some competing
technology.  In this case the placement service will ensure that
launch happens on a host which is capable of *some* kind of memory
encryption, and then later at spawn-time the virt driver can figure
out which technology can be used to deliver it.  This work item covers
the latter stage.

Change-Id: I615aa63a4bfe367dc93bf6805ea0038591eebd8b
blueprint: amd-sev-libvirt-support