Amend "Configure max number of volumes to attach" spec
Update the spec to reflect what was actually implemented.

TL;DR is during implementation, we found that the old limit in the
libvirt driver of 26 was a limit on the maximum number of disk
devices allowed to attach to a single instance, including the root
disk (and any other disks). So "volumes" wasn't really correct for
representing what is being limited and the terminology was changed
to "disk devices".

Related to blueprint conf-max-attach-volumes

Change-Id: I3152d0ed64709495ff7f13ff1d75ce62558a8731