Spec: Ussuri: Encrypted Emulated Virtual TPM
There are a class of applications which expect to use a TPM device to
store secrets. In order to run these applications in a virtual machine,
it would be useful to expose a virtual TPM device within the guest.
This spec describes adding flavor/image properties which cause such a
device to be added to the VM.

Blueprint: add-emulated-virtual-tpm

Change-Id: I299903a5f3b3741cb2b2d0271087c263552d4134
Previously-approved: train
Previously-approved: stein