Encrypted Emulated Virtual TPM
There are a class of applications which expect to use a TPM device to
store secrets. In order to run these applications in a virtual machine,
it would be useful to expose a virtual TPM device within the guest.
This spec describes adding flavor/image properties which cause such a
device to be added to the VM.

Change-Id: I3179413b91e528b5b5d86825768fef98c9dae33c
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>