Smart-nic Support Specification
This spec proposes an overall design for smartnic management which
will involve Nova, Neutron and Cyborg changes.

After the discusstion during Wallaby PTG, we have a general agreement
about the overall design. Please refer to:
https://etherpad.opendev.org/p/nova-wallaby-ptg

In this spec, we will introduce basically how to manage a smart-nic
and how to use a smart-nic for a VM.

Co-authored-by: Yongli He<yongli.he@intel.com>

Change-Id: Ia425629ae5222d6fa7485845533231d0c04a989b