Libvirt supports composing cyborg owned vGPU accelerator into domain XML
In order to allow operators to use cyborg to manage the lifecycle of vGPU,
cyborg needs to discover the vGPUs, reports them to placement, and nova virt
driver helps to allocate this cyborg-owned vGPU to an instance.

In the current cyborg-nova interaction [1]_, there is still a small gap in
allocating a cyborg owned vGPU to an instance. This spec proposes to support
composing cyborg owned vGPU into domain XML in nova libvirt driver.

Co-Authored-By: YumengBao <yumeng_bao@yahoo.com>

Part of blueprint cyborg-vgpu-support

Change-Id: Icaf5556417741a27c83c1e0fe5b5fc39fbac48b6