Change error code when vDPA is unsupported
We had specified that we would return HTTP 403 (Forbidden) in the event
that a unsupported operation was requested on an instance with one or
more vDPA interfaces attached. This was chosen due to us already using
this code for unsupported operations with cyborg accelerators. However,
HTTP 403 (Forbidden) isn't a great choice for this, as it suggests
issues with authorization as opposed to a request the server simply
can't support. HTTP 409 (Conflict) isn't much better, but it does avoid
confusion with policy issues and aligns us with two other existing
examples: unsupported operations on instances with vTPM devices and
instances with SEV enabled.

Change-Id: Ia5d77416b497ef04fb915841b8cdf300230ff00c
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>