Integration With Off-path Network Backends
https://blueprints.launchpad.net/nova/+spec/integration-with-off-path-network-backends

Off-path SmartNIC DPUs introduce an architecture change where
network agents responsible for NIC switch configuration and representor
interface plugging run on a separate SoC with its own CPU, memory and
that runs a separate OS kernel. The side-effect of that is that
hypervisor hostnames no longer match SmartNIC DPU hostnames which are
seen by ovs-vswitchd and OVN agents while the existing port binding code
relies on that. The goal of this specification is to introduce changes
necessary to extend the existing hardware offload code to cope with the
hostname mismatch and related design challenges while reusing the rest
of the code. To do that, PCI(e) add-in card tracking is introduced for
boards with unique serial numbers so that it can be used to determine
the correct hostname of a SmartNIC DPU which is responsible for a
particular VF. Additionally, more information is suggested to be passed
in the "binding:profile" during a port update to facilitate representor
port plugging.

Change-Id: I07ef52769da72cde8867f996111b7df4a80e4d79