Repropose "vm-boot-with-unaddressed-port" spec
This was initially proposed during the Train cycle but was never
completed. As stated in the original change, currently nova only allows
to boot a VM if all ports attached:

- Have an IP assigned.
- The IP allocation is deferred.

Nova doesn't allow to spawn a VM with a port without an IP assigned,
however, some kind of applications (e.g.: routers) need a port which may
be receiving traffic for several addresses, already configured in other
devices.

At the moment, neutron permits to create a port, assigned to a network
with the corresponding sub-networks and IP pools, without an IP. This
spec would allow nova to use these ports.

Change-Id: I94b3559e00e22dd1cde5f35a766c15a5b0dce8a9
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>