Repropose: Support Consumer Types in placement
This spec was previously proposed in Train [1] and the corresponding
implementation has been slowly worked on since. The implementation is
nearly completed now having ironed out many wrinkles that were brought
up during code review.

This contains updates to the design which were discussed and agreed
upon during code review with the main difference being that there is no
longer a default consumer type proposed and that a value of ``NULL``
will indicate the absence of a consumer type for a consumer. Along with
this, no online data migration is now needed.

Previously-approved: Train

Story: 2005473
Task: 30554

[1] https://docs.openstack.org/placement/latest/specs/train/approved/2005473-support-consumer-types.html

Change-Id: I41ce8c336e9917de645faf94d5a31898b0fa0c81