Revert "Amend configurable-instance-hostnames to include response changes"
This reverts commit aec9a01d54ce02d24745e80cc9d9635a0eeb7048. This
change proposed renaming ``OS-EXT-SRV-ATTR:hostname`` attribute in
``/servers`` responses to match the ``hostname`` attribute in requests,
which would avoid a mismatch between the server requests and server
responses. However, this mismatch already exists for the ``host`` and
``node`` attributes of the server, which are represented as ``host`` and
``hypervisor_hostname`` in requests (from microversion 2.74) but
``OS-EXT-SRV-ATTR:host`` and ``OS-EXT-SRV-ATTR:hypervisor_hostname`` in
responses. In addition, there are many more fields with the
``OS-EXT-SRV-ATTR`` prefixes and it seems odd to remove it for one field
(and force clients to make changes) without removing the other
extension-based prefixes for both the various ``/servers`` APIs and the
``/flavors`` APIs. It would be better to tackle this in a separate
microversion or potentially not at all [1].

[1] https://etherpad.opendev.org/p/nova-api-cleanup (Item #3, line 46)

Change-Id: I4caff30f2b3cc12f0970874dcbee04e572e8ccc5
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>