Re-propose QoS minimum guaranteed packet rate
Most of the implementation has been landed in Xena in Nova except the
support for nova-manage placement heal_allocations CLI. So the spec is
re-proposed to finish that part up in Yoga.

Note that Neutron did not merged the support in Xena they will do it in
Yoga probably.

blueprint: qos-minimum-guaranteed-packet-rate
Change-Id: Ieb661461cda0f0c6734d78cb5416c484722a2bcb