Late Amendments to the Off-path Backends Spec
This change mainly addresses the issue with reusing the existing
``VNIC_SMARTNIC`` vnic type.

As discussed on IRC, this will run into a conflict with the existing
Ironic use-case with SmartNICs since the remote_managed tag addition to
an InstancePCIRequest is based the usage of ``VNIC_TYPE_SMARTNIC`` in
the current design:

https://meetings.opendev.org/irclogs/%23openstack-nova/%23openstack-nova.2022-02-07.log.html#t2022-02-07T15:49:56

In order to address that, a new ``VNIC_TYPE_REMOTE_MANAGED`` vnic type
is introduced into neutron-lib:

https://review.opendev.org/c/openstack/neutron-lib/+/828174
https://review.opendev.org/c/openstack/neutron-specs/+/828173

Additionally, this change actualizes some of the implementation details
and cleans up references to dependent changes in different projects
(OVN, Libvirt).

Change-Id: Ia8a537a3de08333262410f8ce1d1d0c6ae997e1a