Remove setup.py and setup.cfg
Theses used to be used for building docs in the past and were made
obsolete with change Icc65afd5f6c0e22147b179c16d6a4fa5f47954ce.

A recent release of setuptools 61.0.0 introduced a breaking change [1]
which caused our openstack-tox-docs and openstack-tox-pep8 jobs to fail
with:

  error: Multiple top-level packages discovered in a flat-layout:
    ['specs', 'priorities'].

This removes setup.py and setup.cfg from use in tox.ini to resolve the
issue and they are not needed anyway.

Related-Bug: #1966382

[1] https://github.com/pypa/setuptools/issues/3197

Change-Id: Ib90065a62bdef235e83173c73894c478af0df5fa