Make 'tox -e fast-specs' fail on sphinx warnings
I'm not sure if others will want this but when I made an edit on my
spec and ran 'tox -e fast-specs', my expectation was that it was a fast
way to check whether the edits I made would pass openstack-tox-docs in
the gate. This turned out not to be the case, 'tox -e fast-specs'
showed the "fast-specs: commands succeeded" message but when I uploaded
the change, the openstack-tox-docs job failed because there was a
sphinx parsing error in my spec.

This adds -W to the sphinx-build call to treat warnings as errors to
match 'tox -e docs' and openstack-tox-docs. It also runs sphinx-build
via exec to make the script fail if sphinx-build returns non-zero.

Change-Id: I9de750e7f44a746d4466c522004d0c2a624052e2