Configurable instance domains
Currently, there is no way for a user booting an instance to tell the
instance's guest operating system what is the domain name of the instance.
Using Fully Qualified Domain Names (FQDN) as the instance hostname was
unintentionally allowed for a period of time, but with the merging of [1] that
ability went away. This spec proposes a new API attribute to enable users to
set the domain name of their instances.

Change-Id: Ifa785634b9d20ed076a511619dbed293ae4aa7b0
Implements: blueprint configurable-instance-domains