Revert "Configurable instance domains"
This reverts commit 35aaec78108fabd1961274250a8fa460abed6f90.

Reason for revert: after this was merged there emerged that
other cores had several concerns with the direction of this spec,
Both in terms of is this in the scope of nova and
if it is, is this the correct way to support passing fqdn/dns_domain
info to the gust vs the other abandoned approach.

As such we should follow the reverts-for-retrospective-vetos policy
https://docs.openstack.org/nova/latest/contributor/policies.html#reverts-for-retrospective-vetos
and revert this to give time for those issues to be discussed.

Change-Id: Ib7950d805db2891f270f02aa9b595aacdd219aee