Robustify Compute Node Hostnames backlog spec
This is mostly a brain-dump spec on the topic of compute hosts and
how fragile they are in terms of hostname handling. There has long
been a requirement that computes can NEVER change hostnames, and we
have few tools to even detect the problem before we corrupt the
database if it happens.

Here I have documented some of the things we could do to make that
more robust, should we choose to do so. This is based on a recent
near-catastrophe and thus reflects things that would have avoided
pain in a real scenario.

Per discussion at the PTG, I am adding this as a backlog spec, to be
an overarching guide for multiple smaller specs to provide more
detailed progress towards the goals described here.

Change-Id: I72fa3f605cfcf7c3dd0ff4c791be7df8f19f058b