Use extend volume completion action
This spec proposes to use the os-extend_volume_completion volume action
that has been proposed for Cinder to provide feedback on success or
failure when handling volume-extended external server events.

A previous version of this spec proposed a new API in Nova, but is has
since been rewritten completely.
The old version can be found in Patchset 6 and earlier.

Blueprint: assisted-volume-extend
Change-Id: I1a59cd571dfcec814df09e208979a481d4ea43af