Re-propose spec for ephemeral storage encryption
Previously-approved: Zed, Yoga, Xena, Wallaby

The spec has been updated to reflect the consensus that we will not
expose low-level encryption options to end users for the initial
version because of the security and upgrade implications discussed.
We will instead hard-code sensible defaults for the relevant options.

Related to blueprint ephemeral-storage-encryption

Change-Id: I425bf537413fbb633f9dec068a247883dbf8680d