Review usage of oslo-privsep library on Nova
Nova's usage of the privsep library is too broad. A single global permission
profile with all needed capabilities is defined for all functions that interact
with privsep to use. While this works, it is not the best usage of the library
as functions are getting a set of rights they do not need and thus should not
receive. This spec seeks to fix this situation by defining a more specialized
usage of the library.

Change-Id: I713ddd60275f29f4361a4c6e93b7299f1d31aa85