Re-propose spec for ephemeral storage encryption
Previously-approved: 2023.1, Zed, Yoga, Xena, Wallaby

The spec has been updated to reflect a change needed in the
implementation to support snapshot and shelve for ephemeral encrypted
instances. The encryption secret is needed in order to boot a new
instance from a snapshot of an ephemeral encrypted instance and to
unshelve an ephemeral encrypted instance. So, the spec is updated to
propose an additional flavor extra spec or image property to keep the
encryption secret UUID from the key manager.

Related to blueprint ephemeral-storage-encryption


Change-Id: Ie796abd247a9231d1e126a589f03a81696960b26