Amend spec to add more details around encryption secrets
This removes the mention of a flavor extra spec for indicating the
encryption secret UUID of an encrypted image. The secret is related
only to an image, not a flavor.

This also adds more details about the handling of encryption secrets
when instances with ephemeral encryption are snapshotted, shelved, or
rescued.

Related to blueprint ephemeral-storage-encryption

Change-Id: I3599e2fe59113b1d27ba08019c7cd78f1ab5a725