nova-specs should use upper-constraints
A recent release of python-subunit 1.4.3 caused
stestr to break due to an undeclared transitive dep.
this is fixed in stestr 4.1.0 but it highlighted that
nova-specs is not using uc when installing and as a result
is prone to failures that wont happen in other tox venvs.

This change simply enables the use of upper-constraints
to prevent future breakages like this.

Change-Id: Ic8a19b8426d3d91f82e907ace62f846a9d6c7282