Add openapi spec
We would like to start documenting our APIs in an industry-standard,
machine-readable manner. Doing so opens up many opportunities for both
OpenStack developer and OpenStack users alike, notably the ability to
both auto-generate and auto-validate both client tooling and
documentation alike. Of the many API description languages available,
OpenAPI (fka "Swagger") appears to be the one with both the largest
developer mind share and the one that would be the best fit for
OpenStack due to the existing tooling used in many OpenStack services,
thus we would opt to use this format.

Change-Id: Ia712c98ff4388bbfa72232324899b1f0cd6293fb
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>